64 bit and verbosity fixes, from difx-1.5
git-svn-id: https://svn.atnf.csiro.au/difx@2241 9d4a3555-dc7e-4673-bb7d-d8a0b9a30876

+++ mpifxcorr/trunk/src/datastream.h
@@ -28,6 +28,7 @@
 #include <vector>
 #include <iostream>
 #include <pthread.h>
+#include <stdint.h>
 #include "architecture.h"
 #include "configuration.h"
 
@@ -177,7 +178,7 @@ protected:
   * Attempts to open the next frame by reading data from the open network socket
   * @return 0 on failure, otherwise the framesize in bytes
   */
-  virtual int openframe();
+  virtual uint64_t openframe();
 
  /** 
   * Attempts to open a network socket to read data from
@@ -212,7 +213,7 @@ protected:
   * @param buffersegment The segment of the databuffer that this read will be stored in
   * @param framebytesremaining The number of bytes left in this frame before reading a new frame is necessary
   */
-  virtual void networkToMemory(int buffersegment, int & framebytesremaining);
+  virtual void networkToMemory(int buffersegment, uint64_t & framebytesremaining);
 
  /** 
   * Reads the specified number of bytes from the specified socket into the provided buffer
